[WIP][UN-2682] Move schema loading to ConfigureConnectorModal
Major refactoring to optimize connector configuration flow:

- **ConfigureConnectorModal**:
  - Moved specConfig and isSpecConfigLoading state management from parent
  - Added fetchEndpointConfigSchema function to load JSON schema
  - Schema now loads once when modal opens (prevents duplicate API calls)
  - Can now select and configure connectors without file/folder selection
  - Simplified prop interface by removing redundant props

- **DsSettingsCard**:
  - Removed schema-related state management (now in modal)
  - Removed all legacy connector listing code and unused functions
  - Cleaned up imports and removed unused constants
  - Simplified to focus on connection type selection and modal trigger

This refactoring improves performance and user experience by allowing
connector configuration without requiring file system browsing.
